+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+### Changed
+- Drop addresses in transaction list
+
+### Fixed
 ## [3.8.3] - 2022-06-10
 
 ### Added